🚚 Renamed `TOC` layout component and added to index
