+++ src/components/layouts/index.js
@@ -1,3 +1,4 @@
 export { default as Layout } from './DefaultLayout'
 export { default as HomeHeader } from './HomeHeader'
 export { default as NavBar } from './NavBar'
+export { default as TOC } from './TOC'
