+++ src/templates/doc-navigation-toc.js
@@ -8,7 +8,7 @@ import { Spirit } from '../components/spirit-styles'
 import NavSidebar from '../components/global/navigation-sidebar'
 import FeedbackForm from '../components/global/feedback-form'
 import PrevNext from '../components/global/prev-next'
-import TOC from '../components/layouts/partials/toc'
+import TOC from '../components/layouts/TOC'
 import { MetaData } from '../components/meta'
 import Icon from '../components/global/icon'
 import getMetaImageUrls from '../utils/getMetaImageUrls'
